Some renaming done.
