+++ test/CMakeLists.txt
@@ -32,7 +32,4 @@ set_target_properties(${CURRENT_PROJECT_NAME}_chacha_test PROPERTIES CXX_STANDAR
 cm_test(NAME ${CURRENT_PROJECT_NAME}_hash_test SOURCES ${CURRENT_TEST_SOURCES_DIR}/hash.cpp)
 set_target_properties(${CURRENT_PROJECT_NAME}_hash_test PROPERTIES CXX_STANDARD 17)
 
-cm_test(NAME ${CURRENT_PROJECT_NAME}_hash_based_algebraic_test SOURCES ${CURRENT_TEST_SOURCES_DIR}/hash_based_algebraic.cpp)
-set_target_properties(${CURRENT_PROJECT_NAME}_hash_based_algebraic_test PROPERTIES CXX_STANDARD 17)
-
 include_directories(${CMAKE_WORKSPACE_SOURCES_DIR})
